create READING.md
adding a separate, cordoned off reading list. I don't intend for this repo to become that, especially with another reading list repo out there, but that one is overwhelmingly large at this point. I'd like to have a reading list I'd be willing to consider "mandatory" for new managers, which means it can't be massive/exhaustive.